Got to here so far
Getting a strange error: "Exception in <: #<procedure => is not a real number"
I added print statements in run and moves. The program indicates that it is getting into run and moves once before the exception occurs. My assumption is that there may be something wonky with the moves func. I will investigate.

We are no longer getting compile errors or warnings and the program is running some so as far as I know syntax is okay.